makefile tweaks

+++ Makefile
@@ -1,6 +1,6 @@
 #!make
 GO_CMD = $(shell which go)
-SHELL = /usr/bin/bash
+SHELL = bash
 
 build:
 	GOOS=linux GOARCH=amd64 $(GO_CMD) build -o bin/tzLookupService-linux-amd64 main.go
@@ -17,7 +17,8 @@ build-all:
 clean:
 	rm -fv bin/*
 	rm -frv dist
-	rm -v timezone.snap.db
+	rm -frv release
+	if [ -f timezone.snap.db ]; then rm -v timezone.snap.db; fi
 
 test:
 	$(GO_CMD) test
@@ -27,3 +28,7 @@ create-tz-db:
 	curl -LO -C - https://github.com/evansiroky/timezone-boundary-builder/releases/download/2020a/timezones.geojson.zip
 	unzip timezones.geojson.zip
 	go run cmd/timezone.go -json "dist/combined.json" -db=timezone -type=boltdb
+
+release: create-tz-db build-all
+	mkdir release
+	cp -v bin/*  timezone.snap.db release/
